cloud_function/requirements.txt
functions-framework==3.*
google-cloud-pubsub
faker
fake-useragent